move time codecs outside of test directory
